put cursor movement in it's own script to clean up placetrap a little
