+++ Assets/Scripts/Players/MoveControllerCursor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62d64fb07f6ee1840af9a9c3970d7512
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
